# validator/nym/directory/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a temporarily centralized directory/PKI/metrics API to allow us to get the other Nym node types running. Its functionality will eventually be folded into other parts of Nym.",
        "title": "Nym Directory API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "https://github.com/nymtech/nym-directory/license"
        },
        "version": "0.9.0-dev"
    },
    "paths": {
...
        "/api/mixmining": {
            "post": {
                "description": "Nym network monitor sends packets through the system and checks if they make it. The network monitor then hits this method to report whether the node was up at a given time.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mixmining"
                ],
                "summary": "Lets the network monitor create a new uptime status for a mix",
                "operationId": "addMixStatus",
                "parameters": [
                    {
                        "description": "object",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MixStatus"
                        }
                    }
                ],
                "responses": {
                    "201": {},
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/api/mixmining/{pubkey}/history": {
            "get": {
                "description": "Lists all mixnode statuses for a given node pubkey",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mixmining"
                ],
                "summary": "Lists mixnode activity",
                "operationId": "listMixStatuses",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Mixnode Pubkey",
                        "name": "pubkey",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.MixStatus"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/api/mixmining/{pubkey}/report": {
            "get": {
                "description": "Provides summary uptime statistics for last 5 minutes, day, week, and month",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mixmining"
                ],
                "summary": "Retrieves a summary report of historical mix status",
                "operationId": "getMixStatusReport",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Mixnode Pubkey",
                        "name": "pubkey",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {},
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "models.MixStatus": {
            "type": "object",
            "required": [
                "ipVersion",
                "pubKey",
                "up"
            ],
            "properties": {
                "ipVersion": {
                    "type": "string"
                },
                "pubKey": {
                    "type": "string"
                },
                "up": {
                    "type": "boolean"
                }
            }
        },
...
    }
}